+++ src/ScottPlot4/ScottPlot.Tests/PlotTypes/Signal.cs
@@ -1,6 +1,7 @@
 ﻿using NUnit.Framework;
 using System;
 using System.Collections.Generic;
+using System.Drawing;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -254,6 +255,19 @@ namespace ScottPlotTests.PlotTypes
 
             var plt = new ScottPlot.Plot(400, 300);
             plt.AddSignal(values);
+
+            TestTools.SaveFig(plt);
+        }
+
+        [Test]
+        public void Test_Signal_FillDisable()
+        {
+            // https://github.com/ScottPlot/ScottPlot/issues/2436
+
+            ScottPlot.Plot plt = new(400, 300);
+            var sig = plt.AddSignal(ScottPlot.DataGen.Sin(51));
+            sig.FillAboveAndBelow(Color.Red, Color.Blue);
+            sig.FillDisable();
 
             TestTools.SaveFig(plt);
         }